awq mappings registry
