+++ src/llmcompressor/modifiers/awq/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
 
 from .base import *
+from .mappings import *
